Reactivate the site navigation
I don't like that the whole sitemap / directory tree is shown on the left
hand side. But it seems not so simple to avoid this; I don't know a
solution yet. Therefore - for the time being - I activated the normal
site navigation again. Otherwise it could turn out to be hard to get to
one or the other page.